requirements.txt
docker
gitpython
pandas
sklearn
google-cloud-storage
mlflow